Added a CameraController script.
This script will force a camera (Main Camera usually, but you can set any camera) to follow a specified GameObject with an assigned height and distance from the object. This may be your player, or an enemy, or a even a car, dream big! This script has an option to zoom the camera in and out depending on the speed in which the target is moving. This script also has an option to delay the camera catching up to the target when it is rotating, showing the sides of the object as it rotates. Both features are fully customisable for how long they last and how extreme they are.